My thoughts on this problem
The problem states that all lights are left on and that each switch always controls its corresponding light, meaning switch in room A always controls light A and switch B controls light B. The problem occurs when switches are connected to multiple lights for example Switch A controls light A and B, this means light B is controlled by both switch A and B. In this example if switch A is pressed both lights turn off, but if light B is pressed only light B turns off then if switch A is pressed light A gets turned off and light B gets turned back on this means switch B needs to be clicked again to turn it off. From this example the first case of just pushing switch A is the desired solution that we are trying to find instead of pushing switches ABA.
To solve the problem we try to find a way that sends a signal from a switch to any given light an odd number of times in the minimum number of switch pushes. This is because each light has only 2 states, on or off and all lights start as on. Therefore, if the state changes an odd number of times for any light it is off. To do this we create an number of switches by number of lights matrix with a 1 stored for each connection between a switch and light(an edge) this produces an adjacency matrix. This then allows as to sum each column, and this represents weather we can turn of a given light when we push all of the switches if the number produced is odd. When summing if you remove a row then that switch isn’t pushed therefore we can find the switches needed to push by summing each combination of rows for every column finding the one that needs the least number of rows to produce the largest number of odd summed columns. This method is the method we used for are program because all off the lights are on when we solve. If we wanted to be able to solve the problem with some lights on and others off we would need to also store the state of each light if the some methodology was to be used we could store the state of each light in a matrix with 0 being off and 1 being on then adding the value to the corresponding summed column therefore if the light is on you must push the switch an even number of times to produce an odd sum.
For this specific question, we will be provided with two lines of input, first line defining the lights with letters, and second line a set of edges. For example, circuit:
A B C D
AC AD BD CA
This means switch A is connected to lights A,C and D, B is connected to lights B and D, C is connected to lights C and A and D is connected to light D. Since this problem input format does not supply information on light state, we assume they are all on. We then can produce an matrix with lights values with size equal to the number of lights, and therefore we would have an array, (1,1,1,1) with position 0 indicates the state of light A, and so on. Where 1 shows they are ON.
This means we have four lights in the house. The edge AC means the switch A is controlling C and A, the edge AD controlling A and D, and therefore, in are program we compress this to ACD, meaning switch A is controlling light A, C and D.
Once this information is collected and stored we can start to solve this problem, the basic idea of are solution is that we try to turn off the lights by pressing each combination of switches one by one and write down the states of all the lights, until we have turned off all the lights. Then we need to record two things, the switches that were pressed, and the states of the lights after pressing the switches.
The first step in the program is that we find all the possible combinations of switches we can press ignoring order in which they are pressed. We then simulate each combination of button presses store the number of lights still on in a hash map.
For example, say if we have A B C D, then the premutation of those for switches are (A, B, C, D, AB, AC, AD, BC, BD, CD, ABC, ABD, BCD, ABCD). We don’t have to consider the order as the states of lights after pressing AB and BA is the same.
So, my hash map will look like this after finding the premutation,
Key, value
A, 0
B, 0
C, 0
D, 0
AB, 0
AC, 0
… ..
Then we can test them one by one with the edges supplied and see the state of lights. For example, since we previously stored the edges AC and AD as ACD, we can search for A in the hash map and get that A turns of ACD so the light states stored in the array will become(0,1,0,0) after pressing A and the sum of lights is 0+1+0+0=1. Since one light is still on the sum of lights is 1 and we save this the hash map. Then we reset the light state arrayto its initial state (1,1,1,1) and continue going through each permutation stored in the hash map and record the states. We then search for the best solution in the hash map.
Some cases
Line 1: A B C
Line 2:
This is the most simple case where each of the three lights A B and C with their only connections being themselves therefore to turn off a light all switches connected to a light that is on must be pressed once.
1: A B C
Line 2: AB BC BA
This case has three lights connected in a loop this means if switch A turns off B and A leaving C on. To turn off C switch C must be pressed but this also turns light A back on. Each button press turns the neighbouring light back on therefore one light must always remain on.
Line 1: A B C D
Line 2: AB BC CD DA
This Case also produces a loop but since there is an even number of switches within the loop all lights can be turned off by pressing either AC or BD.